requirements.txt
langchain==0.1.20
langchain-openai==0.1.6
langchain-community==0.0.38
openai>=1.24.0,<2.0.0
duckduckgo-search==5.0.0
python-dotenv==1.0.0
aiohttp==3.9.1

